Abstracted payment method queries into standalone functions.
