+++ imports/plugins/core/payments/server/no-meteor/queries/index.js
@@ -1,0 +1,7 @@
+import availablePaymentMethods from "./availablePaymentMethods";
+import paymentMethods from "./paymentMethods";
+
+export default {
+  availablePaymentMethods,
+  paymentMethods
+};
